refactor: split deploy code evidence into focused modules
Move file indexing and signal loading out of code-evidence.ts into file-index.ts so repository scanning limits, git-aware file discovery, and readable signal creation have a dedicated boundary.

Move TechnicalBaseline loading and normalization into baseline-evidence.ts, and move runtime/dependency evidence derivation into dependency-evidence.ts while preserving the existing DeploymentCodeEvidence schema and public code-evidence facade exports.

Keep shared evidence value helpers in evidence-utils.ts to avoid duplicate ref/value/fingerprint compaction logic as deploy evidence rules continue to grow.

Verified with npm run build, npm run verify:deploy-structured-blockers, and npm run smoke:deploy.
